requirements.txt
streamlit 
langchain 
google-generativeai 
pandas 
datetime
langchain_community
plotly
langchain-google-genai
pdfminer.six
langchain-text-splitters 
chromadb 
nltk